🎨 Structure: make directorys
