+++ package.json
@@ -12,12 +12,10 @@
     "class-validator": "^0.14.1",
     "formik": "^2.4.6",
     "next": "14.2.3",
-    "next": "14.2.3",
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.2.1",
     "react-toastify": "^10.0.5"
-    "yarn": "^1.22.22"
   },
   "devDependencies": {
     "@types/node": "^20",
@@ -25,11 +23,8 @@
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
-    "eslint": "^8",
-    "eslint-config-next": "14.2.3",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
     "typescript": "^5"
   }
 }
